Fixes the break from build https://travis-ci.org/maurerit/abandapart/builds/148951832

+++ aba-industry/industry-calculator/src/main/java/com/aba/industry/web/IndustryRestController.java
@@ -36,7 +36,7 @@ public class IndustryRestController {
                     MediaType
                             .APPLICATION_JSON_VALUE )
     public BuildCalculationResult calculateBuild ( @PathVariable( "systemName" ) String systemName,
-                                                   @PathVariable( "typeId" ) Long typeId )
+                                                   @PathVariable( "typeId" ) Integer typeId )
     {
         return industryCalculationService.calculateBuildCosts( systemName, typeId );
     }
